chore: move network test to workload_test.go
