+++ internal/pkg/template/template_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/aws/copilot-cli/internal/pkg/aws/s3"
 	"github.com/gobuffalo/packd"
 	"github.com/stretchr/testify/require"
-	"gopkg.in/yaml.v3"
 )
 
 func TestTemplate_Read(t *testing.T) {
@@ -180,119 +179,6 @@ func TestTemplate_Parse(t *testing.T) {
 			} else {
 				require.Equal(t, tc.wantedContent, c.String())
 			}
-		})
-	}
-}
-
-func TestTemplate_ParseNetwork(t *testing.T) {
-	type cfn struct {
-		Resources struct {
-			Service struct {
-				Properties struct {
-					NetworkConfiguration map[interface{}]interface{} `yaml:"NetworkConfiguration"`
-				} `yaml:"Properties"`
-			} `yaml:"Service"`
-		} `yaml:"Resources"`
-	}
-
-	testCases := map[string]struct {
-		input *NetworkOpts
-
-		wantedNetworkConfig string
-	}{
-		"should render AWS VPC configuration for public subnets by default": {
-			input: nil,
-			wantedNetworkConfig: `
-  AwsvpcConfiguration:
-    AssignPublicIp: ENABLED
-    Subnets:
-    - Fn::Select:
-      - 0
-      - Fn::Split:
-        - ','
-        - Fn::ImportValue: !Sub '${AppName}-${EnvName}-PublicSubnets'
-    - Fn::Select:
-      - 1
-      - Fn::Split:
-        - ','
-        - Fn::ImportValue: !Sub '${AppName}-${EnvName}-PublicSubnets'
-    SecurityGroups:
-      - Fn::ImportValue: !Sub '${AppName}-${EnvName}-EnvironmentSecurityGroup'
-`,
-		},
-		"should render AWS VPC configuration for private subnets": {
-			input: &NetworkOpts{
-				AssignPublicIP: "DISABLED",
-				SubnetsType:    "PrivateSubnets",
-			},
-			wantedNetworkConfig: `
-  AwsvpcConfiguration:
-    AssignPublicIp: DISABLED
-    Subnets:
-    - Fn::Select:
-      - 0
-      - Fn::Split:
-        - ','
-        - Fn::ImportValue: !Sub '${AppName}-${EnvName}-PrivateSubnets'
-    - Fn::Select:
-      - 1
-      - Fn::Split:
-        - ','
-        - Fn::ImportValue: !Sub '${AppName}-${EnvName}-PrivateSubnets'
-    SecurityGroups:
-      - Fn::ImportValue: !Sub '${AppName}-${EnvName}-EnvironmentSecurityGroup'
-`,
-		},
-		"should render AWS VPC configuration for private subnets with security groups": {
-			input: &NetworkOpts{
-				AssignPublicIP: "DISABLED",
-				SubnetsType:    "PrivateSubnets",
-				SecurityGroups: []string{
-					"sg-1bcf1d5b",
-					"sg-asdasdas",
-				},
-			},
-			wantedNetworkConfig: `
-  AwsvpcConfiguration:
-    AssignPublicIp: DISABLED
-    Subnets:
-    - Fn::Select:
-      - 0
-      - Fn::Split:
-        - ','
-        - Fn::ImportValue: !Sub '${AppName}-${EnvName}-PrivateSubnets'
-    - Fn::Select:
-      - 1
-      - Fn::Split:
-        - ','
-        - Fn::ImportValue: !Sub '${AppName}-${EnvName}-PrivateSubnets'
-    SecurityGroups:
-      - Fn::ImportValue: !Sub '${AppName}-${EnvName}-EnvironmentSecurityGroup'
-      - "sg-1bcf1d5b"
-      - "sg-asdasdas"
-`,
-		},
-	}
-
-	for name, tc := range testCases {
-		t.Run(name, func(t *testing.T) {
-			// GIVEN
-			tpl := New()
-			wanted := make(map[interface{}]interface{})
-			err := yaml.Unmarshal([]byte(tc.wantedNetworkConfig), &wanted)
-			require.NoError(t, err, "unmarshal wanted config")
-
-			// WHEN
-			content, err := tpl.ParseLoadBalancedWebService(WorkloadOpts{
-				Network: tc.input,
-			})
-
-			// THEN
-			require.NoError(t, err, "parse load balanced web service")
-			var actual cfn
-			err = yaml.Unmarshal(content.Bytes(), &actual)
-			require.NoError(t, err, "unmarshal actual config")
-			require.Equal(t, wanted, actual.Resources.Service.Properties.NetworkConfiguration)
 		})
 	}
 }
